+++ src/Audit.HttpClient/Request.cs
@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using System.Net.Http;
+using System.Text.Json.Serialization;
 
 namespace Audit.Http
 {
@@ -9,5 +11,6 @@ namespace Audit.Http
         public string Path { get; set; }
         public Dictionary<string, string> Headers { get; set; }
         public Content Content { get; set; }
+        public Dictionary<string, object> Options { get; set; }
     }
 }